% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_with_NAs.R
\name{pad_with_NAs}
\alias{pad_with_NAs}
\title{Add NAs to a vector}
\usage{
pad_with_NAs(x, n_left = 1, n_right = 1)
}
\arguments{
\item{x}{Vector to which NAs will be added.}

\item{n_left}{Number of NAs to add before x.}

\item{n_right}{Number of NAs to add after x.}
}
\value{
A vector containing x with the requested number of NA values before and after.
}
\description{
Add NAs to a vector
}
\examples{
pad_with_NAs(1:5, n_left = 0, n_right = 3)
pad_with_NAs(c("spider", "mouse", "cat", "dog"), n_left = 1, n_right = 2)

}
