% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnorm_rounded.R
\name{rnorm_rounded}
\alias{rnorm_rounded}
\title{Simulate rounded Gaussian random variates}
\usage{
rnorm_rounded(n, mu = 0, sigma = 1, digits = 0)
}
\arguments{
\item{n}{Number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{mu}{Vector of means.}

\item{sigma}{Vector of standard deviations.}

\item{digits}{Integer indicating the number of decimal places to be used in rounding. Negative values are used to round to a power of ten, so for example \code{digits = -2} rounds to the nearest hundred. (See 'Details' of \code{base::round()}).}
}
\value{
Vector of Gaussian random variates, rounded to to specified number of decimal places.
}
\description{
Simulate rounded Gaussian random variates
}
\examples{
rnorm_rounded(n = 10)
}
